Isolate HTTP requests that are pipelined together.
When several requests are pipelined, tshark combines all of them into one line.
This script separates pipelined requests and print a new line for each of them.

tshark's output is ambiguous when referer or cookie header is not present
for all requests. In such cases we assign the existing X headers to last
X requests (this seemed to be the case in our limited analysis of the corresponding
packets).

When several requests are pipelined, we are certain that they are sent to the
same host. So we do not misattribute cookies or referrer headers to wrong hosts
even if we don't assign the headers correctly.